src/main/java/com/netzme/controller/PersonController.java
package com.netzme.controller;

import com.netzme.service.PersonService;
import com.netzme.dto.PersonDTO;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/person")
public class PersonController {

    @Autowired
    private PersonService personService;

    @GetMapping
    public PersonDTO getPerson() {
        return personService.getPerson();
    }
}